requirements.txt
# Gold Macro Trend Model — Dependencies
# Install with: pip install -r requirements.txt

# Data sources
fredapi==0.5.2
yfinance==1.2.0
cot-reports==0.1.3
xlrd==2.0.2
openpyxl==3.1.5

# Data manipulation
pandas==2.3.3
numpy==2.3.4
requests==2.32.5
python-dotenv>=1.0.0

# Machine learning
scikit-learn==1.8.0
lightgbm==4.6.0
scipy==1.17.0

# Visualization
matplotlib==3.10.8
seaborn==0.13.2

# Utilities
tqdm==4.67.1
